refactor: typescript for projectPropContext
