+++ src/features/common/types/projectPropContext.d.ts
@@ -1,3 +1,5 @@
+import { Dispatch, SetStateAction } from 'react';
+
 export interface Coordinates {
     lon: number;
     lat: number;
@@ -41,7 +43,7 @@ export interface Coordinates {
     fixedDefaultTreeCount: number;
   }
   
-  export interface Geometry {
+  export interface GeometryB {
     coordinates?: ((((number)[] | null | number)[] | null)[] | null)[] | null;
     type: string;
   }
@@ -50,7 +52,7 @@ export interface Coordinates {
     timezone: string;
     timezone_type: number;
   }
-  export interface Properties {
+  export interface PropertiesA {
     lastUpdated: LastUpdated;
     name: string;
     description?: string | null;
@@ -58,9 +60,9 @@ export interface Coordinates {
     status: string;
   }
     export interface SitesEntity {
-      geometry: Geometry;
+      geometry: GeometryB;
       type: string;
-      properties: Properties;
+      properties: PropertiesA;
     }
     export interface Address {
       zipCode: string;
@@ -146,7 +148,7 @@ export interface Coordinates {
       mainChallenge: string | null
       motivation: string | null
     }
-    export interface Properties {
+    export interface PropertiesB {
       id: string
       _scope: string
       allowDonations: boolean
@@ -177,43 +179,30 @@ export interface Coordinates {
       metadata: SearchMetadata
       options: any[]
     } 
+    export interface SearchProject  {
+        type: string
+        geometry: GeometryB
+        properties: PropertiesB
+    }
   
     
-  export interface SearchProject  {
-      type: string
-      geometry: Geometry
-      properties: Properties
-  }
   
-  // GeoJson
+  // GeoJson//////////////////////////////////////////////////////////////////////////////////
+  export interface GeometryA {
+    coordinates: number[][][]
+    type: string
+    properties?: PropertiesA
+}
   export interface Feature {
-      geometry: Geometry
+      geometry: GeometryA
       type: string
-      properties: Properties
-  }
-    
-    export interface Geometry {
-      coordinates: any[][][]
-      type: string
-  }
-  
-  export interface Properties {
-      lastUpdated: LastUpdated
-      name: string
-      description?: string
-      id: string
-      status: string
-  }
-  
-  export interface LastUpdated {
-      date: string
-      timezone: string
-      timezone_type: number
+      properties: PropertiesA
   }
   export interface GeoJson {
       type: string
       features: Feature[]
     }
+  
     
   
     // ViewPort
@@ -226,8 +215,6 @@ export interface Coordinates {
       propNames: string[]
       props: Props
     }
-    
-    
     export interface ViewPort {
       width: number
       height: number
@@ -246,9 +233,14 @@ export interface Coordinates {
       transitionInterruption?: number
     }
     
+    
+    
   
     // MapState 
   
+    export interface FillTranslate {
+      stops: [number, number[]][]
+    }
     export interface Paint {
       "background-color"?: string
       "fill-color": any
@@ -266,11 +258,6 @@ export interface Coordinates {
       "text-halo-blur"?: number
       "text-halo-width"?: number
     }
-    
-    export interface FillTranslate {
-      stops: [number, number[]][]
-    }
-    
     export interface Layout {
       visibility?: string
       "line-join"?: string
@@ -308,43 +295,25 @@ export interface Coordinates {
       colorMode: string
       colors: Colors
       boundaries: string
-  }
-  export interface SpatialReference {
+    }
+    export interface SpatialReference {
       wkid: number
       latestWkid?: number
     }
-  export interface ArcgisEditorExtent {
+    export interface ArcgisEditorExtent {
     spatialReference: SpatialReference
     xmin: number
     ymin: number
     xmax: number
     ymax: number
-  }
-    export interface Metadata {
+    }
+    export interface MetadataA{
       arcgisStyleUrl: string
       arcgisOriginalItemTitle: string
       arcgisQuickEditorWarning: boolean
       arcgisQuickEditor: ArcgisQuickEditor
       arcgisEditorExtents: ArcgisEditorExtent[]
       arcgisMinimapVisibility: boolean
-    }
-    
-    
-    export interface Colors {
-      boundaries: string
-    }
-    
-    export interface Layer {
-      id: string
-      type: string
-      paint: Paint
-      layout: Layout
-      showProperties?: boolean
-      source?: string
-      "source-layer"?: string
-      minzoom?: number
-      maxzoom?: number
-      filter?: any[]
     }
     export interface Layer {
       id: string
@@ -376,7 +345,7 @@ export interface Coordinates {
       glyphs?: string
       sources: Sources
       layers: Layer[]
-      metadata?: Metadata
+      metadata?: MetadataA
     }
     
     export interface MapState {
@@ -386,6 +355,8 @@ export interface Coordinates {
       minZoom: number
       maxZoom: number
     }
+    
+    
     
    //rasterData
   
@@ -418,14 +389,14 @@ export interface Coordinates {
   }
   
   export interface RasterData {
-      evi: string | undefined;
-      imagery: Imagery
-   }
+    evi: string | undefined;
+    imagery: Imagery
+  }
     
     
-  // plantLocation
+  // samplePlantLocation
   
-  export interface DeviceLocation {
+  export interface Geometry {
     coordinates: number[]
     type: string
   }
@@ -442,10 +413,10 @@ export interface Coordinates {
     deviceModel: string
     deviceManufacturer: string
     deviceBrand: string
-    deviceLocation: DeviceLocation
+    deviceLocation: Geometry
   }
   
-  export interface Metadata {
+  export interface MetadataB {
     app: App
     public: Public
   }
@@ -456,8 +427,50 @@ export interface Coordinates {
     coordinateIndex: string
     id: string
     status: string
+    created?: string
+    updated?: string
+  }
+  export interface Metadata2 {
+    app: App
+    private: any[]
+    public: any[]
+  }
+  export interface NextMeasurementDate {
+    date: string
+    timezone: string
+    timezone_type: number
+  }
+  export interface Measurements {
+    width: number
+    height: number
   }
   
+  export interface RevisionPeriodicity {
+    subsequentInterval: string
+    discreteIntervals: string[]
+    startDate: string
+  }
+  
+  export interface LastMeasurementDate {
+    date: string
+    timezone: string
+    timezone_type: number
+  }
+  export interface EventDate {
+    date: string
+    timezone: string
+    timezone_type: number
+  }
+  export interface History {
+    image: string
+    statusReason: any
+    created: string
+    eventName: string
+    classification: any
+    eventDate: EventDate
+    measurements: Measurements
+    status: string
+  }
   export interface SamplePlantLocation {
     parent: string
     nextMeasurementDate?: NextMeasurementDate
@@ -478,104 +491,22 @@ export interface Coordinates {
     image: any
     idempotencyKey: string
     profile: string
-    coordinates: Coordinate2[]
+    coordinates: Coordinate[]
     revisionPeriodicity?: RevisionPeriodicity
     scientificSpecies: string
     history: History[]
-    originalGeometry: OriginalGeometry
+    originalGeometry: Geometry
     captureMode: string
     geometry: Geometry
     lastMeasurementDate?: LastMeasurementDate
     captureStatus: string
-    deviceLocation: DeviceLocation3
+    deviceLocation: Geometry
     status: any
   }
   
-  export interface NextMeasurementDate {
-    date: string
-    timezone: string
-    timezone_type: number
-  }
   
-  export interface Metadata2 {
-    app: App2
-    private: any[]
-    public: any[]
-  }
-  export interface DeviceLocation2 {
-    coordinates: number[]
-    type: string
-  }
-  export interface App2 {
-    appVersion: string
-    deviceSystemName: string
-    deviceSystemVersion: string
-    deviceModel: string
-    deviceManufacturer: string
-    deviceBrand: string
-    deviceLocation: DeviceLocation2
-  }
+  //plantLocation
   
-  
-  
-  export interface Measurements {
-    width: number
-    height: number
-  }
-  
-  export interface Coordinate2 {
-    image: string
-    created: string
-    coordinateIndex: string
-    id: string
-    updated: string
-    status: string
-  }
-  
-  export interface RevisionPeriodicity {
-    subsequentInterval: string
-    discreteIntervals: string[]
-    startDate: string
-  }
-  export interface Measurements2 {
-    width: number
-    height: number
-  }
-  export interface EventDate {
-    date: string
-    timezone: string
-    timezone_type: number
-  }
-  export interface History {
-    image: string
-    statusReason: any
-    created: string
-    eventName: string
-    classification: any
-    eventDate: EventDate
-    measurements: Measurements2
-    status: string
-  }
-  export interface OriginalGeometry {
-    coordinates: number[]
-    type: string
-  }
-  
-  export interface Geometry {
-    coordinates: number[]
-    type: string
-  }
-  
-  export interface LastMeasurementDate {
-    date: string
-    timezone: string
-    timezone_type: number
-  }
-  
-  export interface DeviceLocation3 {
-    coordinates: number[]
-    type: string
-  }
   
   export interface PlantedSpecy {
     scientificName: string
@@ -586,29 +517,8 @@ export interface Coordinates {
     id: string
     updated: string
   }
-  
-  export interface OriginalGeometry2 {
-    coordinates: number[][][]
-    type: string
-  }
-  
-  export interface Geometry2 {
-    coordinates: number[][][]
-    type: string
-    properties: Properties
-  }
-  
-  export interface Properties {
-    id: string
-  }
-  
-  export interface DeviceLocation4 {
-    coordinates: number[]
-    type: string
-  }
-  
   export interface PlantLocation {
-    metadata: Metadata
+    metadata: MetadataB
     hid: string
     otherSpecies: null
     description: null
@@ -629,11 +539,11 @@ export interface Coordinates {
     plantProject: string
     plantedSpecies: PlantedSpecy[]
     plantDateStart: null
-    originalGeometry: OriginalGeometry2
+    originalGeometry: GeometryA
     captureMode: string
-    geometry: Geometry2
+    geometry: GeometryA
     captureStatus: string
-    deviceLocation: DeviceLocation4
+    deviceLocation: Geometry
     status: null
   }
   
@@ -641,4 +551,89 @@ export interface Coordinates {
       center: number[];
       zoom: number;
   }
+ //main 
+ export type SetState<T> = Dispatch<SetStateAction<T>>;
   
+  export type ProjectContextInterface = {
+    project: Project | null;
+    setProject: SetState<Project | null>;
+    projects: Project[] | null;
+    setProjects: SetState<Project[] | null>;
+    plantLocations: PlantLocation[] | null;
+    setPlantLocations: React.Dispatch<
+      React.SetStateAction<PlantLocation[] | null>
+    >;
+    showSingleProject: boolean;
+    setShowSingleProject: SetState<boolean>;
+    showProjects: boolean;
+    setShowProjects: SetState<boolean>;
+    searchedProject: SearchProject[];
+    setsearchedProjects: SetState<SearchProject[]>;
+    geoJson: GeoJson | null;
+    setGeoJson: SetState<GeoJson | null>;
+    selectedPl: PlantLocation | null;
+    setSelectedPl: SetState<PlantLocation | null>;
+    zoomLevel: number;
+    setZoomLevel: SetState<number>;
+    siteExists: boolean;
+    setsiteExists: SetState<boolean>;
+    selectedSite: number;
+    setSelectedSite: SetState<number>;
+    isMobile: boolean;
+    setIsMobile: SetState<boolean>;
+    infoRef: {
+      current: null;
+    };
+    exploreContainerRef: React.MutableRefObject<null>;
+    exploreExpanded: boolean;
+    setExploreExpanded: SetState<boolean>;
+    exploreForests: boolean;
+    setExploreForests: SetState<boolean>;
+    explorePotential: boolean;
+    setExplorePotential: SetState<boolean>;
+    exploreDeforestation: boolean;
+    setExploreDeforestation: SetState<boolean>;
+    explorePlanted: boolean;
+    setExplorePlanted: SetState<boolean>;
+    infoExpanded: string | null;
+    setInfoExpanded: SetState<string | null>;
+    openModal: boolean;
+    setModalOpen: SetState<boolean>;
+    filtersOpen: boolean;
+    setFilterOpen: SetState<boolean>;
+    filteredProjects: SearchProject[] | null;
+    setFilteredProjects: React.Dispatch<
+      React.SetStateAction<SearchProject[] | null>
+    >;
+    purpose: string;
+    setPurpose: SetState<string>;
+    mapState: MapState;
+    setMapState: SetState<MapState>;
+    selectedMode: string;
+    setSelectedMode: SetState<string>;
+    viewport: ViewPort;
+    setViewPort: SetState<ViewPort>;
+    exploreProjects: boolean;
+    setExploreProjects: SetState<boolean>;
+    loaded: boolean;
+    setLoaded: SetState<boolean>;
+    mapRef: any;
+    defaultMapCenter: number[];
+    defaultZoom: number;
+    layersSettings: {};
+    setLayersSettings: SetState<{}>;
+    rasterData: RasterData;
+    setRasterData: SetState<RasterData>;
+    satellite: boolean;
+    setSatellite: SetState<boolean>;
+    plIds: string[] | null;
+    setPlIds: SetState<string[] | null>;
+    hoveredPl: PlantLocation | null;
+    setHoveredPl: SetState<PlantLocation | null>;
+    isPolygonMenuOpen: boolean;
+    setIsPolygonMenuOpen: SetState<boolean>;
+    siteViewPort: SiteViewPort | null;
+    setSiteViewPort: SetState<SiteViewPort | null>;
+    plantLocationsLoaded: boolean;
+    setPlantLocationsLoaded: SetState<boolean>;
+  };
